migrate bioauth to typescript
