+++ src/AppNavigator.tsx
@@ -10,7 +10,6 @@ import {useDispatch, useSelector} from 'react-redux'
 import {showErrorDialog, signin} from '../legacy/actions'
 import IndexScreen from '../legacy/components/IndexScreen'
 import MaintenanceScreen from '../legacy/components/MaintenanceScreen'
-import BiometricAuthScreen from '../legacy/components/Send/BiometricAuthScreen'
 import {CONFIG} from '../legacy/config/config'
 import KeyStore from '../legacy/crypto/KeyStore'
 import env from '../legacy/env'
@@ -26,6 +25,7 @@ import {
   isSystemAuthEnabledSelector,
 } from '../legacy/selectors'
 import type {State} from '../legacy/state'
+import {BiometricAuthScreen} from './BiometricAuth'
 import {Boundary} from './components'
 import {CustomPinScreen} from './FirstRun/CustomPinScreen'
 import {FirstRunNavigator} from './FirstRun/FirstRunNavigator'
