+++ src/Catalyst/CatalystNavigator.tsx
@@ -2,10 +2,10 @@ import {createStackNavigator} from '@react-navigation/stack'
 import React from 'react'
 import {useIntl} from 'react-intl'
 
-import BiometricAuthScreen from '../../legacy/components/Send/BiometricAuthScreen'
 import globalMessages from '../../legacy/i18n/global-messages'
 import {defaultNavigationOptions, defaultStackNavigatorOptions} from '../../legacy/navigationOptions'
 import {CATALYST_ROUTES} from '../../legacy/RoutesList'
+import {BiometricAuthScreen} from '../BiometricAuth'
 import {Step1} from './Step1'
 import {Step2} from './Step2'
 import {Step3} from './Step3'
