+++ src/Dashboard/DashboardNavigator.tsx
@@ -4,7 +4,6 @@ import React from 'react'
 import {defineMessages, useIntl} from 'react-intl'
 
 import DelegationConfirmation from '../../legacy/components/Delegation/DelegationConfirmation'
-import BiometricAuthScreen from '../../legacy/components/Send/BiometricAuthScreen'
 import {UI_V2} from '../../legacy/config/config'
 import {isJormungandr} from '../../legacy/config/networks'
 import {
@@ -13,6 +12,7 @@ import {
   jormunNavigationOptions,
 } from '../../legacy/navigationOptions'
 import {SEND_ROUTES, STAKING_CENTER_ROUTES, STAKING_DASHBOARD_ROUTES, WALLET_ROOT_ROUTES} from '../../legacy/RoutesList'
+import {BiometricAuthScreen} from '../BiometricAuth'
 import {SettingsButton} from '../components/Button'
 import {useWalletName} from '../hooks'
 import {useSelectedWallet} from '../SelectedWallet'
