+++ src/Send/SendScreenNavigator.tsx
@@ -4,10 +4,10 @@ import React from 'react'
 import {defineMessages, useIntl} from 'react-intl'
 import {useSelector} from 'react-redux'
 
-import BiometricAuthScreen from '../../legacy/components/Send/BiometricAuthScreen'
 import {defaultNavigationOptions, defaultStackNavigatorOptions} from '../../legacy/navigationOptions'
 import {SEND_ROUTES} from '../../legacy/RoutesList'
 import {tokenBalanceSelector} from '../../legacy/selectors'
+import {BiometricAuthScreen} from '../BiometricAuth'
 import {Boundary} from '../components'
 import {AddressReaderQR} from './AddressReaderQR'
 import {AssetSelectorScreen} from './AssetSelectorScreen'
