+++ src/TxHistory/TxHistoryNavigator.tsx
@@ -5,7 +5,6 @@ import {defineMessages, useIntl} from 'react-intl'
 import {StyleSheet, Text, TouchableOpacity, TouchableOpacityProps} from 'react-native'
 import {useSelector} from 'react-redux'
 
-import BiometricAuthScreen from '../../legacy/components/Send/BiometricAuthScreen'
 import {Button} from '../../legacy/components/UiKit'
 import {UI_V2} from '../../legacy/config/config'
 import {defaultNavigationOptions, defaultStackNavigatorOptions} from '../../legacy/navigationOptions'
@@ -14,6 +13,7 @@ import {tokenBalanceSelector, transactionsInfoSelector} from '../../legacy/selec
 import {COLORS} from '../../legacy/styles/config'
 import {formatDateToSeconds} from '../../legacy/utils/format'
 import iconGear from '../assets/img/icon/gear.png'
+import {BiometricAuthScreen} from '../BiometricAuth'
 import {Boundary, Icon} from '../components'
 import {useWalletName} from '../hooks'
 import {buildOptionsWithDefault, TxHistoryStackParamList, TxHistoryStackRootProps} from '../navigation'
